add custom image downloading functionality
