+++ sciolyid/core.py
@@ -15,12 +15,10 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import asyncio
-import concurrent.futures
 import os
 from functools import partial
 
 import discord
-from git import Repo
 
 import sciolyid.config as config
 from sciolyid.data import GenericError, database, get_category, logger
@@ -152,30 +150,9 @@ async def get_files(item, retries=0):
         logger.info("fetching files")
         logger.info("item: " + str(item))
         if retries < 3:
-            await download_github()
+            await config.options["download_func"]()
             retries += 1
             return await get_files(item, retries)
         else:
             logger.info("More than 3 retries")
             return []
-
-async def download_github():
-    logger.info("syncing github")
-    executor = concurrent.futures.ThreadPoolExecutor(max_workers=3)
-    loop = asyncio.get_event_loop()
-    try:
-        os.listdir(config.options['download_dir'])
-    except FileNotFoundError:
-        logger.info("doesn't exist, cloning")
-        await loop.run_in_executor(executor, _clone)
-        logger.info("done cloning")
-    else:
-        logger.info("exists, pulling")
-        await loop.run_in_executor(executor, _pull)
-        logger.info("done pulling")
-
-def _clone():
-    Repo.clone_from(config.options["github_image_repo_url"], config.options['download_dir'])
-
-def _pull():
-    Repo(config.options['download_dir']).remote("origin").pull()
